+++ src/use-peer-state.tsx
@@ -1,4 +1,4 @@
-import { useLayoutEffect, useState } from 'react';
+import { useLayoutEffect, useState, useRef } from 'react';
 import Peer from 'peerjs';
 
 const usePeerState = <TState extends {}>(
@@ -7,6 +7,9 @@ const usePeerState = <TState extends {}>(
 ): [TState, Function, string, Peer.DataConnection[]] => {
   const [connections, setConnections] = useState<Peer.DataConnection[]>([]);
   const [state, setState] = useState<TState>(initialState);
+  // We useRef to get around useLayoutEffect's closure only having access
+  // to the initial state since we only re-execute it if brokerId changes.
+  const stateRef = useRef<TState>(initialState);
   const [peer, setPeer] = useState<Peer | undefined>(undefined);
   const [brokerId, setBrokerId] = useState(opts.brokerId);
 
@@ -27,7 +30,7 @@ const usePeerState = <TState extends {}>(
 
           // We want to immediately send the newly connected peer the current data.
           conn.on('open', () => {
-            conn.send(state);
+            conn.send(stateRef.current);
           });
         });
       });
@@ -43,6 +46,7 @@ const usePeerState = <TState extends {}>(
     state,
     (newState: TState) => {
       setState(newState);
+      stateRef.current = newState;
       connections.forEach(conn => conn.send(newState));
     },
     brokerId,